refactor: consolidate 11 crates into 6
Pure restructuring (no behavior change): collapse the pure-Rust crates by
dependency weight, keeping the heavy/optional backends separate.

- router + control + master -> quillcache-core (as modules, re-exported at root)
- transfer -> quillcache-store (mod transfer, re-exported)
- gateway -> the root quillcache binary (src/gateway.rs)

Final crates: quillcache (bin), quillcache-core, quillcache-store, and the three
optional backends index-holt / index-rocksdb / cuda — each pulls a heavy toolchain
(Holt, RocksDB's C++, CUDA) kept out of the default build.

Tests crossing the new crate boundaries became store integration tests
(control_dataplane.rs, distributed_pool.rs). 47 tests pass; fmt + clippy clean.
