--- crates/quillcache-control/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "quillcache-control"
-version = "1.0.0-alpha.1"
-edition = "2021"
-description = "Control-plane state and policy service for QuillCache"
-license = "MIT"
-repository = "https://github.com/feichai0017/quillcache"
-
-[dependencies]
-quillcache-core = { path = "../quillcache-core", version = "1.0.0-alpha.1" }
-quillcache-router = { path = "../quillcache-router", version = "1.0.0-alpha.1" }
-serde = { workspace = true }
-thiserror = { workspace = true }
-
-[dev-dependencies]
-quillcache-store = { path = "../quillcache-store", version = "1.0.0-alpha.1" }
