+++ crates/quillcache-core/src/control.rs
@@ -1,10 +1,10 @@
-use quillcache_core::{
+use crate::router::{GreedyStatePlaneRouter, RouteDecision, RouterError, RoutingPolicy};
+use crate::{
     BlockRemovedEvent, BlockStoredEvent, CacheResidency, CacheTier, DataPlane, DataPlaneAction,
     DataPlaneUpdate, EngineEndpoint, EngineRole, ExternalKvBlockKey, IdentityScope, IndexBackend,
     KvBlockKey, KvEvent, KvEventBatch, MemoryIndex, NoDataPlane, RequestShape, ReuseViolation,
     WorkerState,
 };
-use quillcache_router::{GreedyStatePlaneRouter, RouteDecision, RouterError, RoutingPolicy};
 use serde::{Deserialize, Serialize};
 use std::str::FromStr;
 use thiserror::Error;
@@ -28,7 +28,7 @@ pub struct IngestSummary {
 /// What the identity guard did for one request: how many blocks were safe to
 /// reuse, and how many content-matching blocks were *refused* because they are
 /// resident only under a different identity (a naive content cache would have
-/// served them — see `quillcache_core::ReuseViolation`).
+/// served them — see `crate::ReuseViolation`).
 #[derive(Debug, Clone, Default, PartialEq, Eq, Serialize, Deserialize)]
 pub struct ReuseAudit {
     /// Request blocks whose exact identity is resident — safe to reuse.
@@ -623,8 +623,7 @@ impl ControlPlane {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use quillcache_core::{EngineKind, SloTarget};
-    use std::collections::HashMap;
+    use crate::{EngineKind, SloTarget};
 
     fn engine() -> EngineEndpoint {
         EngineEndpoint {
@@ -904,7 +903,7 @@ mod tests {
 
     #[test]
     fn control_plane_data_plane_defaults_to_none_and_is_pluggable() {
-        use quillcache_core::MockDataPlane;
+        use crate::MockDataPlane;
         let control = ControlPlane::new(vec![engine()]);
         // By default there is no tensor data plane (infer from events).
         assert_eq!(control.data_plane().name(), "none");
@@ -958,59 +957,5 @@ mod tests {
             .actions
             .iter()
             .any(|action| action.kind == PlanActionKind::Decode));
-    }
-
-    #[test]
-    fn tiered_data_plane_updates_runtime_residency() {
-        use quillcache_store::{StoreDataPlane, StoreTierConfig};
-        let mut control = ControlPlane::new(vec![engine()]).with_data_plane(Box::new(
-            StoreDataPlane::new(StoreTierConfig {
-                hbm_capacity_bytes: 100,
-                cpu_dram_capacity_bytes: 100,
-                local_ssd_capacity_bytes: 100,
-            }),
-        ));
-        let blocks = ["a", "b", "c", "d"]
-            .iter()
-            .enumerate()
-            .map(|(idx, hash)| {
-                KvBlockKey::new(
-                    "Qwen/Qwen3-0.6B",
-                    "Qwen/Qwen3-0.6B",
-                    "tenant-a",
-                    "root",
-                    *hash,
-                    idx as u32,
-                    64,
-                )
-            })
-            .collect::<Vec<_>>();
-        let request = RequestShape {
-            id: "tiered".to_string(),
-            model_id: "Qwen/Qwen3-0.6B".to_string(),
-            tokenizer_id: "Qwen/Qwen3-0.6B".to_string(),
-            adapter_id: None,
-            tenant_id: "tenant-a".to_string(),
-            session_id: None,
-            blocks,
-            estimated_decode_tokens: 16,
-            slo: SloTarget::default(),
-        };
-
-        let actions = control.observe_placement("vllm-a", &request, 100);
-        assert!(actions
-            .iter()
-            .any(|action| action.kind == quillcache_core::DataPlaneActionKind::Evict));
-        let tiers = control
-            .residency()
-            .snapshot()
-            .into_iter()
-            .map(|entry| (entry.key.block_hash, entry.tier))
-            .collect::<HashMap<_, _>>();
-        assert_eq!(tiers.len(), 3);
-        assert_eq!(tiers.get("d"), Some(&CacheTier::Hbm));
-        assert_eq!(tiers.get("c"), Some(&CacheTier::CpuDram));
-        assert_eq!(tiers.get("b"), Some(&CacheTier::LocalSsd));
-        assert!(!tiers.contains_key("a"));
     }
 }
