--- crates/quillcache-gateway/Cargo.toml
@@ -1,30 +1,0 @@
-[package]
-name = "quillcache-gateway"
-version = "1.0.0-alpha.1"
-edition = "2021"
-description = "OpenAI-compatible gateway and event ingest service for QuillCache"
-license = "MIT"
-repository = "https://github.com/feichai0017/quillcache"
-
-[dependencies]
-axum = { workspace = true }
-quillcache-control = { path = "../quillcache-control", version = "1.0.0-alpha.1" }
-quillcache-core = { path = "../quillcache-core", version = "1.0.0-alpha.1" }
-quillcache-router = { path = "../quillcache-router", version = "1.0.0-alpha.1" }
-quillcache-store = { path = "../quillcache-store", version = "1.0.0-alpha.1" }
-quillcache-index-holt = { path = "../quillcache-index-holt", version = "1.0.0-alpha.1", optional = true }
-quillcache-index-rocksdb = { path = "../quillcache-index-rocksdb", version = "1.0.0-alpha.1", optional = true }
-futures-util = "0.3"
-reqwest = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
-serde_yaml = { workspace = true }
-thiserror = { workspace = true }
-tokio = { workspace = true }
-tracing = { workspace = true }
-
-[features]
-# Back the online control plane with a persistent ART (Holt) residency index.
-holt = ["dep:quillcache-index-holt"]
-# Back it with a RocksDB (LSM) residency index (needs a C++ toolchain).
-rocksdb = ["dep:quillcache-index-rocksdb"]
