--- crates/quillcache-router/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "quillcache-router"
-version = "1.0.0-alpha.1"
-edition = "2021"
-description = "Routing policies for QuillCache"
-license = "MIT"
-repository = "https://github.com/feichai0017/quillcache"
-
-[dependencies]
-quillcache-core = { path = "../quillcache-core", version = "1.0.0-alpha.1" }
-serde = { workspace = true }
-thiserror = { workspace = true }
